+++ src/TimeTracker.App/Binders/SystemDataTimeProviderBinder.cs
@@ -3,7 +3,7 @@ using TimeTracker.App.Infrastructure.Common;
 
 namespace TimeTracker.App.Binders;
 
-internal class SystemDataTimeProviderBinder : BinderBase<SystemDateTimeProvider>
+internal sealed class SystemDataTimeProviderBinder : BinderBase<SystemDateTimeProvider>
 {
     protected override SystemDateTimeProvider GetBoundValue(BindingContext bindingContext) => new();
 }