+++ src/TimeTracker.App/MainOperations.cs
@@ -1,7 +1,7 @@
 ﻿using TimeTracker.App.Application.Activities;
 using TimeTracker.App.Domain.Common;
 using TimeTracker.App.Domain.Extensions;
-using TimeTracker.App.Infrastructure.Common;
+using TimeTracker.App.Infrastructure.Common.Interfaces;
 using TimeTracker.App.Infrastructure.Repositories;
 
 namespace TimeTracker.App;
@@ -10,28 +10,18 @@ internal static class MainOperations
 {
     internal static void CreateNewActivityHandler(
         ActivityRepository activityRepository,
-        SystemDateTimeProvider dateTimeProvider,
+        IDateTimeProvider dateTimeProvider,
         double hours,
         string activity)
     {
-        var today = dateTimeProvider.Today;
-        var created = new CreateActivityCommand(today, hours, activity)
-            .Pipe(activityRepository.CreateTimeRecord);
-        if (created is false)
-        {
-            Console.WriteLine("New Activity not added");
-            return;
-        }
-
-        Console.WriteLine("New Activity created");
-        var activities = new GetAllActivitiesQuery(5, Ordering.Descending)
+        _ = activityRepository.CreateActivity(new CreateActivityCommand(dateTimeProvider.Today, hours, activity));
+        var activityViews = new GetAllActivitiesQuery(5, Ordering.Descending)
             .Pipe(activityRepository.GetAllActivities)
-            .ToList();
+            .Select(ActivityView.From);
 
-        var activityDtos = activities.ConvertAll(ActivityView.From);
-        foreach (var a in activityDtos)
+        foreach (var view in activityViews)
         {
-            Console.WriteLine(a);
+            Console.WriteLine(view);
         }
     }
 }